Plugin Name and base class name changed : CustomPlugin -> DevicePlugin
