+++ Assets/Scripts/DevicePlugins.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba74279ea8b4b030193edeeb41986d9e
+guid: 6e7ddf7828da5a5f5b9115ddf3a3ef30
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
