+++ Assets/Scripts/DevicePlugins/CameraPlugin.cs.meta
