+++ Assets/Scripts/DevicePlugins/DepthCameraPlugin.cs.meta
