+++ Assets/Scripts/DevicePlugins/DevicePlugin.cs.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: b244c0ec4779f9c6db127680d6683769
+guid: 1e96dcbe1f04705dcbb7f22fbda354ba
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 700
+  executionOrder: 0
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
