+++ Assets/Scripts/DevicePlugins/MultiCameraPlugin.cs.meta
